introduce GeoServer interceptor lib and boot application
