+++ shogun-config/src/main/resources/application-base.yml
@@ -35,6 +35,7 @@ spring:
   redis:
     host: localhost
     port: 6379
+    password: pass
 
 keycloak:
   auth-server-url: http://localhost:8000/auth
