+++ shogun-gs-interceptor/shogun-gs-interceptor-boot/src/main/java/de/terrestris/shogun/interceptor/config/InterceptorWebSecurityConfig.java
