+++ shogun-gs-interceptor/shogun-gs-interceptor-lib/src/main/java/de/terrestris/shogun/interceptor/exception/InterceptorException.java
@@ -1,0 +1,21 @@
+package de.terrestris.shogun.interceptor.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+public class InterceptorException extends Exception {
+
+    public InterceptorException(String message) {
+        super(message);
+    }
+
+    public InterceptorException(Throwable cause) {
+        super(cause);
+    }
+
+    public InterceptorException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+}
