+++ shogun-gs-interceptor/shogun-gs-interceptor-lib/src/main/java/de/terrestris/shogun/interceptor/request/WfsRequestInterceptorInterface.java
