+++ shogun-gs-interceptor/shogun-gs-interceptor-lib/src/main/resources/db/migration/V0.2.0__Baseline.sql
