--- shogun-gs-interceptor/src/main/java/de/terrestris/shogun/interceptor/exception/InterceptorException.java
@@ -1,22 +1,0 @@
-package de.terrestris.shogun.interceptor.exception;
-
-public class InterceptorException extends Exception {
-
-    private static final long serialVersionUID = 1L;
-
-    public InterceptorException() {
-    }
-
-    public InterceptorException(String message) {
-        super(message);
-    }
-
-    public InterceptorException(Throwable cause) {
-        super(cause);
-    }
-
-    public InterceptorException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-}
