--- shogun-gs-interceptor/src/main/resources/application.yml
@@ -1,39 +1,0 @@
-spring:
-  profiles:
-    include: base
-    active: base, interceptor
-
----
-spring:
-  profiles: interceptor
-  datasource:
-    url: jdbc:postgresql://localhost:5555/shogun?currentSchema=interceptor
-  flyway:
-    defaultSchema: interceptor
-    schemas: interceptor, public
-
-server:
-  port: 8082
-  servlet:
-    context-path: /shogun-gs-interceptor
-
-interceptor:
-  namespaceBoundUrl: true
-  defaultOwsUrl: http://localhost:1234/geoserver
-  namespaces:
-    - namespace: SHOGUN
-      url: http://localhost:1234/geoserver/SHOGUN
-
-keycloak:
-  auth-server-url: http://localhost:8000/auth
-  realm: SHOGUN
-  resource: shogun-gs-interceptor
-  public-client: true
-  principal-attribute: preferred_username
-
-## TODO: use auth! / Token!
-keycloakauth:
-  username: admin
-  password: shogun
-  master-realm: master
-  admin-client: admin-cli
