--- shogun-gs-interceptor/src/site/site.xml
@@ -1,1 +1,0 @@
-<project name="SHOGun Boot" combine.self="override" />
